consolidate some test code
