+++ cli/commands/genesis/root_test.go
@@ -21,38 +21,17 @@
 package genesis_test
 
 import (
-	"context"
-	"fmt"
 	"testing"
 
-	corestore "cosmossdk.io/core/store"
-	"cosmossdk.io/log"
-	"cosmossdk.io/store"
-	"cosmossdk.io/store/metrics"
-	storetypes "cosmossdk.io/store/types"
-	"github.com/berachain/beacon-kit/chain"
 	"github.com/berachain/beacon-kit/cli/commands/genesis"
 	"github.com/berachain/beacon-kit/config/spec"
 	"github.com/berachain/beacon-kit/consensus-types/types"
-	"github.com/berachain/beacon-kit/log/noop"
-	nodemetrics "github.com/berachain/beacon-kit/node-core/components/metrics"
 	"github.com/berachain/beacon-kit/primitives/bytes"
 	"github.com/berachain/beacon-kit/primitives/common"
 	"github.com/berachain/beacon-kit/primitives/crypto"
-	cryptomocks "github.com/berachain/beacon-kit/primitives/crypto/mocks"
 	"github.com/berachain/beacon-kit/primitives/math"
-	"github.com/berachain/beacon-kit/primitives/transition"
 	"github.com/berachain/beacon-kit/primitives/version"
-	"github.com/berachain/beacon-kit/state-transition/core"
-	"github.com/berachain/beacon-kit/state-transition/core/mocks"
-	statedb "github.com/berachain/beacon-kit/state-transition/core/state"
-	"github.com/berachain/beacon-kit/storage"
-	"github.com/berachain/beacon-kit/storage/beacondb"
-	"github.com/berachain/beacon-kit/storage/db"
-	depositstore "github.com/berachain/beacon-kit/storage/deposit"
-	dbm "github.com/cosmos/cosmos-db"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/stretchr/testify/mock"
+	statetransition "github.com/berachain/beacon-kit/testing/state-transition"
 	"github.com/stretchr/testify/require"
 )
 
@@ -106,7 +85,7 @@ func TestStateTransitionGenesis(t *testing.T) {
 	cs, err := spec.MainnetChainSpec()
 	require.NoError(t, err)
 
-	sp, st, ds, ctx := setupState(t, cs)
+	sp, st, ds, ctx := statetransition.SetupTestState(t, cs)
 	var (
 		genPayloadHeader = new(types.ExecutionPayloadHeader).Empty()
 		genVersion       = version.FromUint32[common.Version](version.Deneb)
@@ -123,88 +102,4 @@ func TestStateTransitionGenesis(t *testing.T) {
 	processorRoot, err := st.GetGenesisValidatorsRoot()
 	require.NoError(t, err)
 	require.Equal(t, expectedValRoot, processorRoot)
-}
-
-func setupState(t *testing.T, cs chain.Spec) (
-	*core.StateProcessor[*transition.Context],
-	*statedb.StateDB,
-	*depositstore.KVStore,
-	*transition.Context,
-) {
-	t.Helper()
-
-	execEngine := mocks.NewExecutionEngine(t)
-
-	mocksSigner := &cryptomocks.BLSSigner{}
-	mocksSigner.On(
-		"VerifySignature",
-		mock.Anything, mock.Anything, mock.Anything,
-	).Return(nil)
-
-	dummyProposerAddr := []byte{0xff}
-
-	kvStore, depositStore, err := initTestStores()
-	require.NoError(t, err)
-	beaconState := statedb.NewBeaconStateFromDB(kvStore, cs)
-
-	sp := core.NewStateProcessor[*transition.Context](
-		noop.NewLogger[any](),
-		cs,
-		execEngine,
-		depositStore,
-		mocksSigner,
-		func(bytes.B48) ([]byte, error) {
-			return dummyProposerAddr, nil
-		},
-		nodemetrics.NewNoOpTelemetrySink(),
-	)
-
-	ctx := &transition.Context{
-		SkipPayloadVerification: true,
-		SkipValidateResult:      true,
-		ProposerAddress:         dummyProposerAddr,
-	}
-
-	return sp, beaconState, depositStore, ctx
-}
-
-var testStoreKey = storetypes.NewKVStoreKey("state-transition-tests")
-
-type testKVStoreService struct {
-	ctx sdk.Context
-}
-
-func (kvs *testKVStoreService) OpenKVStore(context.Context) corestore.KVStore {
-	//nolint:contextcheck // fine with tests
-	store := sdk.UnwrapSDKContext(kvs.ctx).KVStore(testStoreKey)
-	return storage.NewKVStore(store)
-}
-
-func initTestStores() (*beacondb.KVStore, *depositstore.KVStore, error) {
-	db, err := db.OpenDB("", dbm.MemDBBackend)
-	if err != nil {
-		return nil, nil, fmt.Errorf("failed opening mem db: %w", err)
-	}
-	var (
-		nopLog        = log.NewNopLogger()
-		noopCloseFunc = func() error { return nil }
-		nopMetrics    = metrics.NewNoOpMetrics()
-	)
-
-	cms := store.NewCommitMultiStore(
-		db,
-		nopLog,
-		nopMetrics,
-	)
-
-	cms.MountStoreWithDB(testStoreKey, storetypes.StoreTypeIAVL, nil)
-	if err = cms.LoadLatestVersion(); err != nil {
-		return nil, nil, fmt.Errorf("failed to load latest version: %w", err)
-	}
-
-	ctx := sdk.NewContext(cms, true, nopLog)
-	testStoreService := &testKVStoreService{ctx: ctx}
-	return beacondb.New(testStoreService),
-		depositstore.NewStore(testStoreService, noopCloseFunc, nopLog),
-		nil
 }
